further note implementation, working on journalist next
update to some classes to store a number instead of the actual enum for subjects since we are storing the enum as an int in the database. As well as a custom exception specifically for this enum.

Added a way to subscribe to subject.

New server responses added for previously mentioned functions.

edits to websocket service to fit new functions.

changes to repository classes now that there is a fleshed out database design.

edit to docker-compose